requirements/base.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.2.5
idna==2.9
markovify==0.8.0
multidict==4.7.6
Unidecode==1.1.1
websockets==6.0
yarl==1.4.2
tinydb==4.1.1
